LineageDE

Fits NB-GAM to test alternative vs null model for collection of genes. Performs likelihood ratio test and returns p-value based off of chi-squared distribution (ori.pVal) If samples of pseudotime are provided, will also perform permutation test and give an empirical p-value (emp.pVal) and a parametric p-value (par.pVal) from fitting null distribution with a gamma distribution

Usage

LineageDE(
  genes,
  counts,
  ori.pseudotimes,
  conditions,
  ori.weights = NULL,
  experiments = NULL,
  linCompare = 1,
  k = 3,
  samples = NULL,
  sub.pseudotimes = NULL,
  sub.weights = NULL,
  parallel = FALSE,
  n_cores = NULL,
  output = NULL,
  filePath = NULL,
  save.name = NULL,
  fit_time = 100,
  seed = 123
)

Arguments

genes

vector of names or indices specifying genes to be tested

counts

matrix (size = cells by genes) of raw (integer) count values of gene expression

ori.pseudotimes

matrix (size = cells by lineages) of pseudotime values

conditions

factor (length = cells) specifying condition to which each cell belongs

ori.weights

matrix (size = cells by lineages) of weights from pseudotime analysis if applicable (otherwise NULL)

experiments

factor (length = cells) specifying replicate/batch to which each cell belongs if applicable (otherwise NULL)

linCompare

integer specifying which lineage (index) to test (defaults to 1 - only one lineage provided in pseudotimes)

k

integer specifying number of knots in NB-GAM fit (default = 3)

samples

matrix (size = cells by num_samples) specifying indices cells utilized in each sample of pseudotime if applicable (otherwise NULL)

sub.pseudotimes

matrix (size = cells by lineages by num_samples) samples of pseudotime values if applicable (otherwise NULL)

sub.weights

matrix (size = cells by lineages by num_samples) samples of pseudotime weights if applicable (otherwise NULL)

parallel

boolean specifying whether or not to run in parallel (requires BiocParallel)

n_cores

integer specifying number of cores to use if running in parallel

output

integer specifying how often to output results to csv file (number of genes). Defaults to NULL (output after all genes completed).

filePath

string specifying path of directory where files are to be saved (Defaults to current working directory)

save.name

string specifying prefix name of saved csv files (Defaults to "DE_results"). Specify "None" if do not want to save output to csv.

fit_time

time permitted (in seconds) for NB-GAM fits for single gene (defaults to 100)

seed

set random seed

Value

dataframe with LRT, ori.pVal, (emp.pVal), (par.pVal) for each gene tested